pipelines/post-dissemination/vector-db/etl/requirements.txt
google-cloud-bigquery
python-dotenv
openai
numpy
faiss-cpu
openai
psycopg2-binary
SQLAlchemy
tiktoken